fixed warnings about Q_OBJECT and thousand separators had to move some classes to header files
