+++ src/statusbar.h
@@ -33,6 +33,8 @@ private:
 
 class StatusBar : public QStatusBar
 {
+  Q_OBJECT;
+
 public:
   StatusBar(QWidget* parent=nullptr);
